backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
sqlmodel==0.0.22
sqlalchemy==2.0.36
aiosqlite==0.20.0
jinja2==3.1.4
